resize badge list contents, add popover
